chore: proxy menu to allow single selection only
